перераспределение функций по модулям
